+++ project/script/tool/git/git.ps1
@@ -1,39 +1,33 @@
 #Requires -RunAsAdministrator
 
-<# Install Git ---------------------------------------------------------------#>
-function Git_Install() {
-	Write-Host -NoNewline " ==> Install "
-	Write-Host -ForegroundColor Yellow "Git"
-	If (Get-Command git -ErrorAction SilentlyContinue) {
-		Write-Host -NoNewline "  => Upgrade "
-		Write-Host -ForegroundColor DarkYellow "Git"
-		choco upgrade -y git
-	} Else {
-		choco install -y git
-	}
-}
-
 <# Configure Git -------------------------------------------------------------#>
-function Git_Config() {
+function Git_Config_Global() {
 	Write-Host -NoNewline " ==> Config "
-	Write-Host -ForegroundColor Yellow "Git"
+	Write-Host -NoNewline -ForegroundColor Yellow "Git"
+	Write-Host " (Global)"
 
 	# color.ui
+	Write-Host "  => color.ui"
 	git config --global color.ui auto
 
 	# help.autocorrect
+	Write-Host "  => help.autocorrect"
 	git config --global help.autocorrect 1
 
 	# pull.rebase
+	Write-Host "  => pull.rebase"
 	git config --global pull.rebase true
 
 	# push.default
+	Write-Host "  => push.default"
 	git config --global push.default simple
 
 	# rerere.enabled
+	Write-Host "  => rerere.enabled"
 	git config --global rerere.enabled true
 
 	# merge.tool
+	Write-Host "  => merge.tool"
 	If (Get-Command meld -ErrorAction SilentlyContinue) {
 		git config --global merge.tool meld
 		git config --global mergetool.meld.cmd "'C:/Program Files (x86)/Meld/Meld.exe' `$LOCAL `$REMOTE `$BASE `$MERGED"
@@ -45,6 +39,7 @@ function Git_Config() {
 	}
 
 	# diff.tool
+	Write-Host "  => diff.tool"
 	If (Get-Command meld -ErrorAction SilentlyContinue) {
 		git config --global diff.tool meld
 		git config --global difftool.meld.cmd "'C:/Program Files (x86)/Meld/Meld.exe' `$LOCAL `$REMOTE"
@@ -56,6 +51,19 @@ function Git_Config() {
 	}
 }
 
-<# Run Script ----------------------------------------------------------------#>
-Git_Install
-Git_Config
+function Git_Config_Local_HoYa() {
+	Write-Host -NoNewline " ==> Config "
+	Write-Host -NoNewline -ForegroundColor Yellow "Git"
+	Write-Host " (Local)"
+	Write-Host -NoNewline "  => Config "
+	Write-Host -NoNewline -ForegroundColor DarkYellow "HoYa"
+	Write-Host ": llHoYall <hoya128@gmail.com>"
+
+	# user.name
+	Write-Host "  => user.name"
+	git config --local user.name "llHoYall"
+
+	# user.email
+	Write-Host "  => user.email"
+	git config --local user.email "hoya128@gmail.com"
+}