+++ project/CHANGELOG.md
@@ -1,20 +1,52 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [v0.0.1] - 2020-02-28
+## [v0.2.4] - 2020-05-18
 ### Added
-* Project creation
-* Project configuration
+* Q-Dir on Windows
 
-## [v0.0.2] - 2020-03-01
+## [v0.2.3] - 2020-05-18
 ### Added
-* Chocolatey on Windows
+* pyenv on Windows
 
-## [v0.1.0] - 2020-04-24
-### Changed
-* Integrated chocolatey architecture
+## [v0.2.2] - 2020-05-17
+### Added
+* VSCode on Windows
+
+## [v0.2.1] - 2020-05-13
+### Added
+* Microsoft Windows Terminal on Windows
+
+## [v0.2.0] - 2020-05-03
+### Fixed
+* Add script files to executable file
+
+## [v0.1.4] - 2020-04-30
+### Added
+* Powershell on Windows
+
+## [v0.1.3] - 2020-04-27
+### Fixed
+* Chocolatey script on Windows
+
+## [v0.1.2] - 2020-04-27
+### Added
+* git on Windows
 
 ## [v0.1.1] - 2020-04-25
 ### Added
 * Firefox on Windows
 * Chrome on Windows
+
+## [v0.1.0] - 2020-04-24
+### Changed
+* Integrated chocolatey architecture
+
+## [v0.0.2] - 2020-03-01
+### Added
+* Chocolatey on Windows
+
+## [v0.0.1] - 2020-02-28
+### Added
+* Project creation
+* Project configuration